Initial creation of the Basis type
Supports representation of the applicable basis to allow the definition of a Blade object over that basis. This will allow us to reason about the algebraic operations applied to multivectors defined as a set of Blades over a given Basis.

Currently the implementation only has built in support for the standard 8-Dimensional linear space used by the 3D Vector Space Grassman Algebra.